change: rename modules and exports
