+++ package.json
@@ -2,15 +2,20 @@
   "name": "@mgn901/mgn901-utils-ts",
   "version": "0.6.0",
   "description": "",
-  "main": "commonjs/index.js",
-  "types": "commonjs/index.d.ts",
-  "files": ["commonjs", "README.md"],
+  "files": [
+    "esmodules",
+    "README.md",
+    "LICENSE"
+  ],
+  "exports": {
+    "./*": "./esmodules/*.js"
+  },
   "scripts": {
     "prepare": "husky",
     "format": "biome check --write --config-path=./ .",
     "lint": "tsc --noEmit && biome check --config-path=./ .",
-    "build": "tsc -p tsconfig.json",
-    "test": "jest --passWithNoTests"
+    "test": "jest --passWithNoTests",
+    "build": "tsc -p tsconfig.build.json"
   },
   "repository": {
     "type": "git",
@@ -27,6 +32,7 @@
     "@jest/globals": "^29.7.0",
     "@swc/core": "^1.11.18",
     "@swc/jest": "^0.2.37",
+    "@types/node": "^22.14.0",
     "husky": "^9.1.7",
     "jest": "^29.7.0",
     "lint-staged": "^15.5.0",
