--- src/index.ts
@@ -1,30 +1,0 @@
-export {
-  dateToUnixTime,
-  dateToUnixTimeMillis,
-  isUnixTime,
-  isUnixTimeMillis,
-  UnixTime,
-  UnixTimeMillis,
-  unixTimeMillisToDate,
-  unixTimeMillisToUnixTime,
-  unixTimeToDate,
-  unixTimeToUnixTimeMillis,
-} from './Date';
-
-export { KeyValueList } from './KeyValueList';
-
-export { Nominal } from './Nominal';
-
-export { PartiallyPartial } from './PartiallyPartial';
-
-export {
-  isBoolean,
-  isNumber,
-  isBigint,
-  isString,
-  isSymbol,
-  isObject,
-  isFunction,
-  isRecord,
-  isTypedArray,
-} from './utilsForTypeGuard';
