+++ src/partially-partial.type.ts
