+++ src/type-guard-utils.ts
